feat(ui5-table-selection-multi): headerSelector property added (#11711)
* feat(ui5-table-selection-multi): headerSelector property added

- provides an enumeration property to switch in the multi-select case
between a select all and a clear all option
(in future further variants might be added).

Fixes: #5873

* feat(ui5-table-selection-multi): headerSelector property added

- provides an enumeration property to switch in the multi-select case
between a select all and a clear all option
(in future further variants might be added).

Fixes: #5873

* feat(ui5-table-selection-multi): headerSelector property added

- provides an enumeration property to switch in the multi-select case
between a select all and a clear all option
(in future further variants might be added).
- CSS intentation fix

Fixes: #5873

* feat(ui5-table-selection-multi): headerSelector property added

- provides an enumeration property to switch in the multi-select case
between a select all and a clear all option
(in future further variants might be added).
- CSS intentation fix
- Tooltip of the clear-all icon should stay in disabled state

Fixes: #5873