fix(ui5-toolbar): fix overflow layout processing in dialogs (#11894)
There was a problem with the ui5-toolbar component when used inside a dialog. The toolbar's overflow layout was not being processed correctly after the dialog was opened, leading to issues with item visibility and layout.

This tracks for a similar scenario, which triggers the processOverflowLayout method of the toolbar. This ensures that the toolbar's layout is recalculated correctly after the dialog is opened, allowing the overflow items to be displayed properly.

Fixes: #11771