fix(ui5-file-uploader): react on focus out only when needed (#11989)
The ui5-tokenizer should be expanded when the focus is on the ui5-file-uploader
and it should be collapsed when the focus is out of the ui5-file-uploader.

Issue:
- The ui5-tokenizer's expand property got updated on focus out of the ui5-tokenizer or the native
type file filed, or even on focus out of a token.

Solution:
- Now the ui5-tokenizer's expand property is getting updated only on focus out
of the ui5-file uploader. The ui5-tokenizer expand property won't change If any of the ui5-file-uploader
internal elements hosts the focus.

Additionally:
- The file size limit playground sample was improved to use the valueState property
and the valueStateMessage slot instead of ui5-messagestrip component instance.