fix(ui5-color-picker): ensure RGB and HSL values are within limits (#11915)
Previously if an user was about to type a RGB or HSL value in the input fields, that were out of boundaries (from 0 to 255 for all RGB fields, 0 to 360 for Hue, 0 to 100(%) for Saturation and Light in HSL), the input fields became disabled, which led to bad UX.

With this change we now 'normalise' the values that are out of the bounds (mentioned above) to the closest one.

For RGB:

values above 255 –> e.g 275 are normalised to 255;
values below 0 –> e.g -20 are normalised to 0;
For HSL:

Hue:

values above 360 –> e.g 375 are normalised to 360;
values below 0 –> e.g -20 are normlised to 0;
Saturation and Light:

values above 100 –> e.g. 120 are normalised to 100;
values below 0 –> e.g. -20 are normalised