fix(ui5-segmented-button): fix action cancellation with Shift or Esc (#12176)
There is an accessibility-related method for cancelling select actions for button-like controls: when the `Space` key is used for selection, and it is hold down, and meanwhile `Shift` or `Escape` is pressed before the `Space` release, on releasing the `Space` the selection action must not be triggered.

<img width="974" height="94" alt="image" src="https://github.com/user-attachments/assets/3514fa1e-5d57-4e8b-94d8-b094c8d5fd51" />

Now this accessibility feature will be implemented for `<ui5-segmentrd-button>` component.


**JIRA: BGSOFUIBALKAN-9029, case 6**