refactor(ui5-dynamic-side-content): replace resize handler with container queries (#12054)
Currently, every component that has responsive paddings or visualisation based on a specific breakpoint apply class / attribute to the corresponding size and this class / attribute is used to style the component. Usually this class / attribute rely on the component width, where width is observed by resize handler in order to detect changes.

With current PR we want to start to use container queries. Container queries enable applying of styles to an element based on the size of the element's container.