feat(ui5-view-settings-dialog): add group option (#11950)
This feature introduces a new `Group` option to the `ui5-view-settings-dialog` allowing users to select a group of options and in what way they should be ordered, similar to the `Sort` option.

To use the Group option, we introduced a `ui5-group-item` that can be slotted in the `groupItems` slot of a `ui5-view-settings-dialog`.

Fixes: #11815